delete event implemented

+++ client/src/components/personalCalendar.js
@@ -153,15 +153,25 @@ function PersonalCalendar() {
     setName('');
   }
 
-  async function deleteEvent(eventId) {
-    const response = await axios.delete(
-      'http://localhost:8282/appointments/delete',
-      {
-        params: {
-          eventId,
+  async function deleteEvent(eventId, eventUsersId) {
+    if (eventUsersId.length <= 2) {
+      await axios.delete(
+        'http://localhost:8282/appointments/delete',
+        {
+          params: {
+            eventId,
+          },
         },
-      },
-    );
+      );
+    } else {
+      await axios.patch(
+        'http://localhost:8282/appointments/remove_user',
+        {
+          eventId,
+          userId,
+        },
+      );
+    }
     getAppointments();
   }
 
@@ -229,10 +239,11 @@ function PersonalCalendar() {
                   </span>
                 ))}
                 <button
+                  className="delete-button"
                   type="submit"
                   onClick={() => {
                     // eslint-disable-next-line no-underscore-dangle
-                    deleteEvent(appointment._id);
+                    deleteEvent(appointment._id, appointment.user_id);
                   }}
                 >
                   delete
